requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.64
botocore==1.40.64
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.120.4
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.7.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
paho-mqtt==2.1.0
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
postgrest==2.24.0
propcache==0.4.1
psutil==7.1.3
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.24.0
requests==2.32.5
s3transfer==0.14.0
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.231
ultralytics-thop==2.0.18
urllib3==1.26.20
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
